pyproject.toml
[project]
name = "py-arch-lab"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13,<3.14"
dependencies = [
    # Web Framework
    "fastapi",
    "granian[reload]",
    "uvloop",
    "sse-starlette>=3.0.2",
    # Database
    "scylla-driver>=3.29.2", # ScyllaDB driver (provides 'cassandra' module)
    # Data Validation
    "pydantic",
    "pydantic-settings",
    "email-validator>=2.3.0",
    # Messaging & Streaming
    "quixstreams[protobuf]>=3.23.0",
    "grpcio-tools>=1.0.0",
    # Cache
    "redis>=6.4.0",
    "async-lru>=2.0.5",
    # Security
    "pyjwt>=2.10.1",
    "bcrypt>=5.0.0",
    # Utilities
    "uuid-v7>=1.0.0",
    "orjson>=3.11.3",
    "anyio>=4.10.0",
    "attrs>=24.0.0",
    "dependency-injector>=4.48.2",
    "loguru>=0.7.3",
    # Monitoring
    "prometheus-client>=0.20.0",
    "prometheus-fastapi-instrumentator>=7.0.0",
    "opentelemetry-api>=1.30.0",
    "opentelemetry-sdk>=1.30.0",
    "opentelemetry-instrumentation-fastapi>=0.51b0",
    "opentelemetry-instrumentation-redis>=0.51b0",
    "opentelemetry-exporter-otlp>=1.30.0",
    "lz4>=4.4.4",
    "uuid-utils>=0.11.1",
]

[dependency-groups]
dev = [
    # Testing
    "pytest",
    "pytest-asyncio",
    "pytest-bdd-ng>=2.3.1",
    "pytest-xdist>=3.8.0",
    "httpx",

    # Code Quality
    "pyright",
    "ruff",
    "pre-commit>=4.3.0",
    "pyrefly>=0.35.0",

    # Deployment & Infrastructure (CDK)
    "aws-cdk-lib>=2.219.0",
    "constructs>=10.4.2",
    "aws-cdk-aws-apigatewayv2-alpha>=2.114.1a0",
    "aws-cdk-aws-apigatewayv2-integrations-alpha>=2.114.1a0",
    "pyyaml>=6.0.3",

    # Additional development tools
    "uvicorn>=0.37.0",
]
test = ["filelock>=3.19.1"]

[tool.ruff]
line-length = 100
indent-width = 4
target-version = "py313"
exclude = ["script/"]

[tool.ruff.lint]
select = [
    "E", # pycodestyle errors
    "F", # pyflakes
    "B", # flake8-bugbear
    "I", # isort import sorting
    "W", # pycodestyle warnings
]
# If warning hints are annoying while coding, check their type and disable them
# They'll still be fixed when you save or run 'ruff check --fix .'
ignore = [
    "I001", # unsorted-imports 
    "E201", # ignore whitespace after '(' warning
    "E202", # ignore whitespace before ')' warning
    "E241", # allow multiple spaces for alignment
    "E272", # allow multiple spaces before keyword
    "E231", # missing whitespace after comma
    "W291", # trailing whitespace
    "W293", # line contains whitespace
    "E501", # ignore line too long warning
    "D107", # ignore missing docstring in __init__
    "C901", # too complex
    "B904", # Within an `except` clause, raise exceptions with `raise ... from err` or `raise ... from None` to distinguish them from errors in exception handling
    "B008", # Do not perform function call in argument defaults (needed for FastAPI Depends)
    "E722",
]
unfixable = ["F401", "F841"] # unused-import, unused-variable 

[tool.ruff.lint.isort]
order-by-type = true
force-sort-within-sections = true
force-single-line = false
lines-after-imports = 2
combine-as-imports = true

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
skip-magic-trailing-comma = false # preserve trailing commas

[tool.pytest.ini_options]
python_files = ["test.py", "test_*.py", "*_test.py"]
pythonpath = ["."]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
markers = [
    "unit: unit",
    "integration: integration",
    "smoke: Fast smoke tests covering critical API endpoints",
    "cdk: CDK infrastructure tests",
    "e2e",
    "asyncio: mark test as async",
    "bdd: BDD scenario test",
    "api: API test",
    "infra: infra test",
    "entity: Entity test",
    "integration: Integration test",
    "compensating_transaction: Tests for SAGA compensating transaction pattern",
    "skip: mark test to be skipped",
]
filterwarnings = [
    "ignore::DeprecationWarning:pydantic._internal._config",
    "ignore::pytest.PytestReturnNotNoneWarning",
    "ignore::DeprecationWarning:gherkin.gherkin_line",
]
testpaths = ["test"]
addopts = "-v --tb=short --strict-markers"

[tool.pyright]

# some checking need to 
typeCheckingMode = "off"
reportMissingModuleSource = false
reportMissingImports = false
reportInvalidTypeForm = false

[tool.pyrefly]
search-path = [".", "test", ".venv/lib/python3.13/site-packages"]
site-package-path = [".venv/lib/python3.13/site-packages"]
# Only check src/ and test/ for fast type checking, excluding CDK-related directories
project-includes = ["src", "test"]
# Exclude both deployment/ and test/deployment/ (CDK has slow JSII bindings)
project-excludes = ["deployment/", "test/deployment/"]

[tool.pyrefly.errors]
missing-attribute = false
missing-argument = false
unexpected-keyword = false


[tool.uv]
package = false
